ajout de methode de recupération des 3 premiers scores en base
