+++ backend/routes/ScoreRoutes.js
@@ -4,9 +4,9 @@ const ScoreController = require('../controllers/ScoreController');
 
 router.post('/scores', ScoreController.createScore);
 router.get('/scores', ScoreController.getAllScores);
+router.get('/scores/top', ScoreController.getTopScores);
 router.get('/scores/:id', ScoreController.getScoreById);
 router.get('/scores/player/:id', ScoreController.getScoresByPlayerId);
-router.get('/scores/top', ScoreController.getTopScores);
 //router.put('/scores/:id', ScoreController.updateScore);
 //router.delete('/scores/:id', ScoreController.deleteScore);
