+++ database.js
@@ -3,7 +3,7 @@ const { Sequelize } = require('sequelize');
 const sequelize = new Sequelize({
     dialect: 'sqlite',
     storage: './database.sqlite' ,
-    //logging: false // Assurez-vous que le chemin vers votre base de données est correct
+    logging: console.log, // Active les logs SQL dans la console
 });
 
 module.exports = sequelize;
