+++ package.json
@@ -40,6 +40,7 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
+    "sequelize-cli": "^6.6.2",
     "typescript": "~5.5.2"
   }
 }
